+++ README.md
@@ -1,8 +1,8 @@
 ![](https://img.shields.io/badge/Microverse-blueviolet)
 
-# leaderboard
+# Leaderboard
 
-> This web application is a simple Leaderboard. Where the user can type his name along with his score.
+> This website allows users to type in their names and scores. Both name & score are sent to an API(leaderboard). Data is also retrieved from the API.
 
 ## Built With
 
@@ -39,7 +39,7 @@ Run `npm start` for development. Run `npm run build` for production.
 
 - GitHub: [@NatiG25](https://github.com/NatiG25)
 - Twitter: [@NatiG87702270](https://twitter.com/NatiG87702270)
-- LinkedIn: [@Temesgen](https://linkedin.com/in/temesgen-g-gorgis-0910a6229 )
+- LinkedIn: [@Temesgen G/gorgis](https://linkedin.com/in/temesgen-g-gorgis-0910a6229 )
 
 ## 🤝 Contributing
