+++ src/index.html
@@ -24,7 +24,7 @@
       <section class="addScore">
         <h2>Add your score</h2>
             <form>
-            <input class="name" type="text" placeholder="Your name" required/> <br>
+            <input class="user" type="text" placeholder="Your name" required/> <br>
             <input class="score" type="text" placeholder="Your score" required /><br>
             <button class="submitBtn" type="submit">Submit</button>
             </form>
